+++ doc/Recon_combinations.md
@@ -7,9 +7,10 @@ Below is a table of all methods, and what components they work with.
 
 
 
-| Method | Identity | Gradient |     | TotalVariation (TV)  | IndicatorBox |
-| ------ | -------- | -------- | --- | --- | ------------ |
-| FBP    |          |          |     |     |              |
-| FDK    |          |          |     |     |              |
-| CGLS   | ✔        | ✔        |     |     |              |
-| SIRT   | ✔        |          |     | ✔?   | ✔            |
+| Method | Identity | Gradient |     | TotalVariation (TV) | IndicatorBox |     | LeastSquares |
+| ------ | -------- | -------- | --- | ------------------- | ------------ | --- | ------------ |
+| FBP    |          |          |     |                     |              |     |              |
+| FDK    |          |          |     |                     |              |     |              |
+| CGLS   | ✔        | ✔        |     |                     |              |     |              |
+| SIRT   | ✔        |          |     | ✔?                  | ✔            |     |              |
+| FISTA  | -        | -        |     | -                   | -            |     | -            |
