+++ webct/blueprints/app/templates/tags/cpu.html.j2
@@ -1,0 +1,1 @@
+<sl-tag variant="warning" pill size="x-small" slot="suffix">CPU</sl-tag>
