+++ webct/blueprints/app/templates/tags/gpu.html.j2
@@ -1,0 +1,1 @@
+<sl-tag variant="success" pill size="x-small" slot="suffix">GPU</sl-tag>
