+++ webct/blueprints/app/templates/tags/slow.html.j2
@@ -1,0 +1,1 @@
+<sl-tag variant="warning" pill size="x-small" slot="suffix">Slow</sl-tag>
