+++ webct/blueprints/reconstruction/templates/tab.reconstruction.html.j2
@@ -13,15 +13,38 @@
 	<div>
 		<sl-select label="Reconstruction Algorithm" value="FDK" id="selectReconstruction">
 			<sl-menu-label>Non-Iterative</sl-menu-label>
-			{# Span elements do not show up when selected.#}
-			<sl-menu-item value="FDK">FDK<span><sl-tag variant="success" pill size="small">GPU</sl-tag> Feldkamp, David & Kress</span></sl-menu-item>
-			<sl-menu-item value="FBP">FBP<span><sl-tag variant="success" pill size="small">GPU</sl-tag> Filtered Back Projection</span></sl-menu-item>
+
+			{# Span elements do not show up when selected,#}
+			{# they make for good descriptive text. #}
+			<sl-menu-item value="FDK">FDK
+				<span> Feldkamp, David & Kress</span>
+				{% include "./tags/gpu.html.j2" %}
+			</sl-menu-item>
+
+			<sl-menu-item value="FBP">FBP
+				<span> Filtered Back Projection</span>
+				{% include "./tags/gpu.html.j2" %}
+			</sl-menu-item>
+
 			<sl-divider></sl-divider>
 			<sl-menu-label>Iterative Methods</sl-menu-label>
-			<sl-menu-item value="CGLS">CGLS<span><sl-tag variant="warning" pill size="small">CPU</sl-tag> Conjugate Gradient Least Squares</span></sl-menu-item>
-			<sl-menu-item value="SIRT">SIRT<span><sl-tag variant="warning" pill size="small">CPU</sl-tag> Simultaneous Iterative Reconstructive Technique</span></sl-menu-item>
-			<sl-menu-item value="MLEM" disabled>MLEM<span><sl-tag variant="success" pill size="small">GPU</sl-tag> Maximum Likelihood Estimation Method</span></sl-menu-item>
+
+			<sl-menu-item value="CGLS">CGLS
+				<span> Conjugate Gradient Least Squares</span>
+				{% include "./tags/cpu.html.j2" %}
+			</sl-menu-item>
+
+			<sl-menu-item value="SIRT">SIRT
+				<span> Simultaneous Iterative Reconstructive Technique</span>
+				{% include "./tags/cpu.html.j2" %}
+			</sl-menu-item>
+
+			<sl-menu-item value="MLEM" disabled>MLEM
+				<span>Maximum Likelihood Estimation Method</span>
+				{% include "./tags/cpu.html.j2" %}
+			</sl-menu-item>
 		</sl-select>
+
 		<sl-select label="Reconstruction Quality" value="medium" id="selectReconQuality" advanced>
 			<sl-menu-item value="0">High
 				<sl-tag slot="suffix" variant="danger" pill size="small">Memory Intensive</sl-tag>
@@ -76,20 +99,28 @@
 <div class="group" id="settingsConstraint">
 	<div>
 		<sl-select label="Reconstruction Constraint" id="selectConOperator" class="" value="box">
-			<sl-menu-item value="box">Indicator Box<span><br>Upper & Lower Bounds</span></sl-menu-item>
-			<sl-menu-item value="tv">Total Variation</sl-menu-item>
+			<sl-menu-item value="box">Indicator Box
+				<span><br>Upper & Lower Bounds</span>
+			</sl-menu-item>
+			<sl-menu-item value="tv">Total Variation
+				{% include "./tags/slow.html.j2" %}
+			</sl-menu-item>
 		</sl-select>
-		<div class="check-property">
+
+		<div class="check-property" advanced>
 			<sl-checkbox id="checkboxConUpper"></sl-checkbox>
 			<sl-input type="number" id="inputConUpper" label="Upper Constraint" value="5"></sl-input>
 		</div>
+
 		<div class="check-property">
 			<sl-checkbox id="checkboxConLower"></sl-checkbox>
 			<sl-input type="number" id="inputConLower" label="Lower Constraint" value="0"></sl-input>
 		</div>
-		<sl-input id="inputConTVIterations" type="number" label="Iterations" step=1 min=1 advanced></sl-input>
-		<sl-checkbox id="checkboxConTVIsotropic" advanced>Isotropic</sl-checkbox>
-		<sl-input type="number" label="Tolerance" step=1 min=1 value="1" id="inputConTVTolerance" advanced><span slot="suffix">e-6</span></sl-input>
+
+		<sl-input id="inputConTVIterations" type="number" label="Iterations" step=1 min=1 value="10" disabled></sl-input>
+		<sl-input type="number" label="Alpha" step=0.01 min=0 value="0.1" id="inputConTVAlpha" advanced disabled></sl-input>
+		<sl-checkbox id="checkboxConTVIsotropic" checked advanced disabled>Isotropic</sl-checkbox>
+		<sl-input type="number" label="Tolerance" step=1 min=1 value="1" id="inputConTVTolerance" advanced disabled><span slot="suffix">e-6</span></sl-input>
 	</div>
 	<button></button>
 </div>
@@ -101,8 +132,8 @@
 			<sl-menu-item value="identity">Identity Operator</sl-menu-item>
 			<sl-menu-item value="gradient">Gradient Operator</sl-menu-item>
 		</sl-select>
-		<sl-input type="number" label="Alpha" value="0.1" id="inputTikAlpha"></sl-input>
-		<sl-select label="Boundary Condition" advanced value="Neumann" id="selectTikBoundary">
+		<sl-input type="number" label="Alpha" value="0.1" disabled id="inputTikAlpha"></sl-input>
+		<sl-select label="Boundary Condition" advanced disabled value="Neumann" id="selectTikBoundary">
 			<sl-menu-item value="Neumann">Neumann</sl-menu-item>
 			<sl-menu-item value="Periodic">Periodic</sl-menu-item>
 		</sl-select>
