+++ webct/blueprints/samples/static/scss/models.scss
@@ -148,9 +148,6 @@ form.materialForm > sl-divider {
 	grid-column-end: 3;
 }
 
-.hidden {
-	display: none;
-}
 
 div.mixture:not(.hidden) {
 	display: grid;
